refactor: Decouple, improve SoC
