+++ api/src/config/constants.ts
@@ -1,5 +1,3 @@
-import { HttpError } from "../utils/error";
-
 export const API_PREFIX = "/api";
 
 //
@@ -49,64 +47,3 @@ export const GENRES: ReadonlyArray<{ id: number; name: string }> = [
   { id: 29, name: "Vocal" },
   { id: 30, name: "World Music" },
 ];
-
-//
-
-export const HTTP_ERROR_MESSAGES: ReadonlyArray<string> = [
-  "The requested page does not exist",
-  "Something went wrong",
-];
-export const HTTP_ERRORS = {
-  "400": "BadRequest",
-  "401": "Unauthorized",
-  "402": "PaymentRequired",
-  "403": "Forbidden",
-  "404": "NotFound",
-  "405": "MethodNotAllowed",
-  "406": "NotAcceptable",
-  "407": "ProxyAuthenticationRequired",
-  "408": "RequestTimeout",
-  "409": "Conflict",
-  "410": "Gone",
-  "411": "LengthRequired",
-  "412": "PreconditionFailed",
-  "413": "PayloadTooLarge",
-  "414": "URITooLong",
-  "415": "UnsupportedMediaType",
-  "416": "RangeNotSatisfiable",
-  "417": "ExpectationFailed",
-  "418": "ImATeapot",
-  "421": "MisdirectedRequest",
-  "422": "UnprocessableEntity",
-  "423": "Locked",
-  "424": "FailedDependency",
-  "425": "UnorderedCollection",
-  "426": "UpgradeRequired",
-  "428": "PreconditionRequired",
-  "429": "TooManyRequests",
-  "431": "RequestHeaderFieldsTooLarge",
-  "451": "UnavailableForLegalReasons",
-  "500": "InternalServerError",
-  "501": "NotImplemented",
-  "502": "BadGateway",
-  "503": "ServiceUnavailable",
-  "504": "GatewayTimeout",
-  "505": "HTTPVersionNotSupported",
-  "506": "VariantAlsoNegotiates",
-  "507": "InsufficientStorage",
-  "508": "LoopDetected",
-  "509": "BandwidthLimitExceeded",
-  "510": "NotExtended",
-  "511": "NetworkAuthenticationRequired",
-} as const;
-export type HttpErrorCodes = keyof typeof HTTP_ERRORS;
-export type HttpErrorNames = (typeof HTTP_ERRORS)[HttpErrorCodes];
-export type HttpErrorMessages = (typeof HTTP_ERROR_MESSAGES)[number];
-
-//
-
-export const DATABASE_ERROR_CODES: {
-  [key: string]: { httpStatusCode: number; response: HttpError };
-} = {
-  "23505": { httpStatusCode: 409, response: new HttpError({ code: "409" }) },
-};
