--- api/src/controllers/lib/stats/index.ts
@@ -1,30 +1,0 @@
-import { Request, Response, NextFunction } from "express";
-import * as statsModel from "../../../models/lib/stats";
-
-export async function gerTracksCountByGenre(
-  req: Request<unknown, unknown, unknown, { excluded: number[] }>,
-  res: Response<{ results: { id: number; name: string; count: number }[] }>,
-  next: NextFunction,
-) {
-  try {
-    res.json({
-      results: await statsModel.queries.countTracksByGenre(req.query.excluded),
-    });
-  } catch (err) {
-    next(err);
-  }
-}
-
-export async function getTracksCountByYear(
-  req: Request<unknown, unknown, unknown, { excluded: number[] }>,
-  res: Response<{ results: { name: string; count: number }[] }>,
-  next: NextFunction,
-) {
-  try {
-    res.json({
-      results: await statsModel.queries.countTracksByYear(req.query.excluded),
-    });
-  } catch (err) {
-    next(err);
-  }
-}
