--- api/src/controllers/processes/index.ts
@@ -1,2 +1,0 @@
-export * as seeding from "./seeding";
-export * as validation from "./validation";
