+++ api/src/controllers/tracks/validation-schemas.ts
@@ -1,0 +1,10 @@
+import Joi from "joi";
+
+import { FILTER_CONDITIONS } from "../../config/constants";
+import { Filter } from "../../types";
+
+export const schemaFilter = Joi.object<Filter>({
+  name: Joi.string().valid("year", "genre"),
+  condition: Joi.string().valid(...FILTER_CONDITIONS),
+  value: Joi.alternatives(Joi.number(), Joi.array().items(Joi.number())),
+});
