+++ api/src/loaders/index.ts
@@ -1,0 +1,8 @@
+import { expressLoader } from "./express-loader";
+
+export * from "./express-loader";
+
+export const appLoader = {
+  // ... Add more loaders here ...
+  expressApp: expressLoader(),
+};
