+++ api/src/middlewares/handle-404-error.ts
@@ -1,6 +1,6 @@
 import util from "util";
 import { Request, Response, NextFunction } from "express";
-import { HttpError } from "../utils/error";
+import { HttpError } from "../utils";
 
 export function handle404Error(
   req: Request,
