--- api/src/models/lib/index.ts
@@ -1,1 +1,0 @@
-export * as stats from "./stats";
