--- api/src/models/lib/stats/index.ts
@@ -1,1 +1,0 @@
-export * as queries from "./queries";
