--- api/src/models/processes/index.ts
@@ -1,1 +1,0 @@
-export * as queries from "./queries";
