--- api/src/models/track/index.ts
@@ -1,3 +1,0 @@
-export * as queries from "./queries";
-export * as validationSchemas from "./validation-schemas";
-export * as validator from "./validator";
