+++ api/src/processes/seed.ts
@@ -1,5 +1,5 @@
-import { traverseDirs } from "../utils/utilities";
-import * as trackModel from "../models/track";
+import { traverseDirs } from "../utils";
+import { trackService } from "../services/tracks";
 import { GENRES } from "../config/constants";
 import { ProcessMessage } from "../types";
 
@@ -8,8 +8,8 @@ const libPath = process.argv[2];
 (async () => {
   if (!process.send) throw new Error("process.send is undefined");
 
-  await trackModel.queries.createGenres(GENRES);
-  await traverseDirs(libPath, trackModel.queries.create);
+  await trackService.createGenres(GENRES);
+  await traverseDirs(libPath, trackService.create);
 
   process.send!({ name: "seeding", status: "success" } as ProcessMessage);
 })();
